Remove unused VariableLister WatchLengthListener
Modify its test to apply no to PartyNightWatch.
